Fix lmbench build issue (#528)
lmbench3 build fails while applying one of the hunks from the local
patch maintained in lkp. The reason was that the same hunk from the
patch has been upstreamed in the source
(https://github.com/intel/lmbench.git) which was conflicting. Hence
generated a patch after removal of the conflicting hunk.


Reviewed-and-tested-by: Srikanth Aithal <srikanth.aithal@amd.com>
Signed-off-by: Suneeth D <Suneeth.D@amd.com>