repo/linux: add Brendan Jackman's asi/next branch
This lives on my personal Github repository. This is a kitchen-sink repo
so deny everything by default with branch_denylist and then just add in
the one specific branch we care about.

Note this branch is based on linux-next, if that's an issue please let
me know.

I'm setting this branch up at the suggestion of Dave Hansen:

https://lore.kernel.org/all/5a97031d-33d4-4710-ab5a-7d8947936038@intel.com/