repo/linux: Add lyude-cki-zone
Now that I'm doing changes to core locking primitives, it's become pretty
obvious I need a way to be able to test stuff with CKI on-demand without
spamming the mailing list. So, add my own repo.

I don't expect to need this all the time, so I've limited this to only test
branches starting with "the-cki-zone".

Signed-off-by: Lyude Paul <lyude@redhat.com>