lib/yaml.rb: use UTF-8 encoding to read json file
Sometimes LANG env is C which impacts the load_json() to results in

	/var/lib/gems/3.3.0/gems/json-2.12.2/lib/json/common.rb:338:in `encode': "<E2>" on US-ASCII (Encoding::InvalidByteSequenceError)
        from /var/lib/gems/3.3.0/gems/json-2.12.2/lib/json/common.rb:338:in `parse'
        from /var/lib/gems/3.3.0/gems/json-2.12.2/lib/json/common.rb:338:in `parse'
        from /var/lib/gems/3.3.0/gems/json-2.12.2/lib/json/common.rb:852:in `load'

The "<E2>" part of the error specifically points to the byte 0xE2,
which is the first byte of a common UTF-8 multi-byte sequence. This
confirms it's a UTF-8 character being misinterpreted. Also JSON files
are almost universally UTF-8. Thus explictly tell File.read to use
UTF-8 to read.

Signed-off-by: Philip Li <philip.li@intel.com>