filters/need_kconfig.rb: treat y and m are exchangeable
The job should not be filtered when kconfig is required to be m
but built as y (or vice versa) if the job doesn't have logic to
test specific behavior of module or builtin. Also the test script
need be able to handle both m and y.

This is true for most tests, but need further tuning to do this
per different tests.

Signed-off-by: Philip Li <philip.li@intel.com>