jobs: disable 'resched' and 'yield' tests in stress-ng
this is per comments and suggestion from PeterZ and Chenyu

in below Link, PeterZ comments the 'resched' test as:
IIRC this benchmark is literally something like running yield() in a
loop, so boohoo. I'll look when there's a real workload that has a
problem.

Chenyu suggests to disable 'yield' test, too

Link: https://lore.kernel.org/all/20251212101244.GN3911114@noisy.programming.kicks-ass.net/
Suggested-by: Chen Yu <yu.c.chen@intel.com>
Signed-off-by: Oliver Sang <oliver.sang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>